refactor: update encryptedExportType filename
