+++ .github/whitelist-capital-letters.txt
@@ -40,7 +40,6 @@
 ./libs/common/src/misc/serviceUtils.spec.ts
 ./libs/common/src/enums/kdfType.ts
 ./libs/common/src/enums/fileUploadType.ts
-./libs/common/src/enums/encryptedExportType.ts
 ./libs/common/src/enums/linkedIdType.ts
 ./libs/common/src/enums/encryptionType.ts
 ./libs/common/src/enums/htmlStorageLocation.ts
