+++ apps/web/src/app/tools/import-export/export.component.ts
@@ -12,7 +12,7 @@ import { LogService } from "@bitwarden/common/abstractions/log.service";
 import { PlatformUtilsService } from "@bitwarden/common/abstractions/platformUtils.service";
 import { UserVerificationService } from "@bitwarden/common/abstractions/userVerification/userVerification.service.abstraction";
 import { PolicyService } from "@bitwarden/common/admin-console/abstractions/policy/policy.service.abstraction";
-import { EncryptedExportType } from "@bitwarden/common/enums/encryptedExportType";
+import { EncryptedExportType } from "@bitwarden/common/enums/encrypted-export-type";
 
 import { UserVerificationPromptComponent } from "../../components/user-verification-prompt.component";
